+++ test/index.test.ts
@@ -91,4 +91,4 @@ describe('next-test-api-route', () => {
             });
         });
     });
-})
+});